# Replace `SpecsSnapshot` with lazier `SpecsPaths`