Digesting is a much heavier operation than finding file paths. It requires more IO resources, and has worse cache semantics because we care about file content rather than only file names. Hence the `Paths` type: https://github.com/pantsbuild/pants/pull/10741

This change allows target-less goals to only digest files they actually care about, e.g. `update-build-files` now only digests BUILD files.

Before:

```
❯ hyperfine -w 1 -r 5 './pants --no-pantsd update-build-files --check --no-fmt ::'
Benchmark 1: ./pants --no-pantsd update-build-files --check --no-fmt ::
  Time (mean ± σ):      5.238 s ±  0.052 s    [User: 4.159 s, System: 0.541 s]
  Range (min … max):    5.195 s …  5.324 s    5 runs
```

After:

```
❯ hyperfine -w 1 -r 5 './pants --no-pantsd update-build-files --check --no-fmt ::'
Benchmark 1: ./pants --no-pantsd update-build-files --check --no-fmt ::
  Time (mean ± σ):      5.100 s ±  0.070 s    [User: 3.888 s, System: 0.480 s]
  Range (min … max):    5.044 s …  5.211 s    5 runs
```

[ci skip-rust]
[ci skip-build-wheels]